Move protocol, transport fixtures into conftest
